+++ rails_generators/fume/simple_theme/simple_theme_gen.rb
@@ -1,9 +1,0 @@
-class SimpleThemeGen < Fume::Gen::Base
-
-  def manifest
-    record do |m|
-      cp_files m, File.join("public", "themes", "**", "*")
-      cp_files m, File.join("public", "stylesheets", "*.css")
-    end
-  end
-end