+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.7.0:
   - add action_attr_accessor method
   - depend rails_i18n
+  - rename simple\_theme to simple_nav and rewirte